+++ src/app/mypage/_components/recipe-detail/index.ts
@@ -1,0 +1,1 @@
+export { RecipeDetail } from './recipe-detail';
