+++ src/app/mypage/_components/sub-text/index.ts
@@ -1,0 +1,1 @@
+export { SubText } from './sub-text';
